Updated main.py, began doing the gain, is stuck
What the plan is:

- def iterminmax(): Iterate through every node of the tree containing the possible moves, create temporal player list and super list, get possible moves and their gain, add them as a branch to the node, call this function again until the depth goes over diffficulty*2

When at max depth (difficulty*2), we should check if the move is from the human, and make it select the one with the best gain for the human. When at an AI step, check which one has the highest gain for the AI thus far, pick that one, and go up another depth.
So at the end, we should only have the best AI move.